refactored MARS and MARS_gui
Created common library MARS_queue that can be called by either to run MARS
